python/03_uv_packages/packages/summarizer/pyproject.toml
[project]
name = "summarizer"
version = "0.1.0"
description = "Example of how to call Google's Gemini model"
readme = "README.md"
authors = [
    { name = "Alexander Serdyukov", email = "alexander.serdyukov@gmail.com" }
]
requires-python = ">=3.12"
dependencies = ["google-genai",]

# [project.scripts]
# summarizer = "summarizer:main"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"
